+++ app/src/main/java/com/wasingun/seller_lounge/extensions/ViewExtensions.kt
@@ -3,7 +3,7 @@ package com.wasingun.seller_lounge.extensions
 import android.view.View
 import androidx.annotation.StringRes
 import com.google.android.material.snackbar.Snackbar
-import com.wasingun.seller_lounge.util.Constants
+import com.wasingun.seller_lounge.constants.Constants
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
